Fix Bug 1222631: Need Promise Polyfill
Updated version of FFO to one that includes the promise polyfill bundled in the ffo.js file.

This is the last version of ffo.js that bundles the polyfill, after this version it is handled as a dependancy. After this upgrade we should see if bower is as smart about dependancies as we think it is.