Fix bug 1226642: Migrate Prism to Bower
Before this change, we used the Prism website to create a custom build
of the library. The website generated two minified files, prism.css and
prism.js, which we checked into the repository.

This change uses Bower to cherry-pick the individual components of the
library that we need. Pipeline then builds the components into complete
bundles like the ones we had before. The result is the same, but this
approach helps us to be more clear about what version of Prism we use
and what we include in our custom build.

As with other libraries managed by Bower, the library code is checked
into the repository but the minified bundles are not.