Fix bug 1237072 - Fix issue with whitenoise and settings
whitenoise needs DJANGO_SETTINGS_MODULE to be set in the environment
otherwise it raises an ImproperlyConfigured exception.

This adds some code to kuma/wsgi.py which only affects dev
environments where it'll set the variable to "settings_local" if it's
not in the environment. That's in line with what we're doing already
elsewhere.