Fix PyPI link; clarify location of .env
The repository root is mounted in the Vagrant VM as ~/src/ . The
problem here is that the entire dotenv section talks about things but
only refers to the path in the Vagrant VM. If you're editing .env
because you can't get Vagrant to work, it's not clear where the .env
file is.

Tried to clarify it a bit, but the wording is still awkward.