Fix bug 1244853: Nix languages.json
This nixes kuma/languages.json and the thing that was using it.

At best, kuma/languages.json was a copy of the file in product-details
which we now pull down during deploys. At worst, it's an out-of-date
copy of the file in product-details which means that
settings.LANGUAGES and kuma/core/apps.py:CoreConfig.language_mapping()
could be out of sync causing who knows what mischief and mayhem.