Fix bug 1244308: Return 404 content if user can't restore the document
Previously, MDN would return an HTTP 404 with an HTML document that
has some stuff on it and lets people restore the document. That's not
great because somehow those pages are getting indexed or otherwise
making search bots grumpy.

Instead of doing that, this changes MDN to return a straight-up
regular 404 page unless the user can restore documents in which case
they get the HTML page that lets them restore documents.