Fix bug 1247447: Fix bootstrapping problem with product-details
If you just built a dev environment and it was pristine and didn't
have any of the product-details files, then you were in a bit of a
pickle because in order to get product-details files, it'd run through
the settings code which required the product-details files which you
didn't have because that's the very thing you were trying to get. Ugh.

This fixes that by moving some code around such that LOCALES and
LANGUAGES get built together (LANGUAGES requires LOCALES anyhow) and
if the product-details files aren't there, it prints a warning and
rolls its eyes, but then returns empty structures. This is ok because
it only happens if there are no product-details files and that should
only be the case if you haven't gotten the product-details, yet.