Fix bug 1237661: Fix wonky admin
This fixes the wonky admin problem where the language of the admin would
change between refreshes.

What's going on is that Django 1.8 introduced a fix to
trans_real.activate where if you pass in '', it does nothing and
returns. The prefixer will set the locale to '' for /admin/, so then
the locale-aware middleware calls translation.activate('') and
suddenly we're all standing in a pool of goo in the middle ages with
flaming arrows aimed at us.

This fixes that.