bug 1255609 - Enable training on false positives
Add additional data to wiki's DocumentSpamAttempt:
* data - JSON-serialized data sent to Akismet
* review - Staff review status (default 'Needs Review')
* reviewer - Staff reviewer
* reviewed - Review date

Old DocumentSpamAttempts do not have this additional data, so their
status is set to "Review Unavailable".

Update the Django admin for DocumentSpamAttempt so that it works better
as a dashboard for second-guessing Akismet:
* Limit the length of title and slug on the list display
* Allow filtering by "Needs Review", to treat the admin as a review
  backlog.
* Add an editable dropdown to quickly review multiple edits from known
  users
* When a review of "Spam" is chosen, it confims Akismet's choice, and no
  futher API call is needed.
* When a review of "Ham / False Positive" is chosen, submit the
  original data to Akismet's submit_ham API endpoint. On failure, revert
  to unreviewed status so it can be tried again.

Add a management command and task to drop submission data for old
records, saving database resources and removing unneeded PII.

Update the anonymization scripts to drop submission data.