bug 1259173 - Rewrite RevisionForm Tests
The RevisionForm is used in different ways in different views. Rewrite the
tests to setup the forms correctly for English edits, new pages, new
translations, and editing translations. Also, since the Akismet payload will
change quite a bit, especially the comment_content, test more of the payload
for each scenario.