bug 957802 - Remove cached folder for NPM
vagrant-cachier was being used to share the VM's /root/.npm folder
between virtual machines. This was causing issues around permissions,
which appear in the ansible output as messages like:

Error: Attempt to unlock /usr/lib/node_modules/bower-installer, which hasn't been locked

This appears to be an issue with reading or writing /root/.npm/_locks,
and interpreting permissions problems as the need to lock or unlock the
file.

In any case, disabling the cache allows the ansible build to globally
install npm packages.