bug 1272791 - Encode attachment data for JS
Data about attachments are assigned to a JavaScript variable so that
they are available to CKEditor. The JSON encoder does not escape
characters for this context, allowing user-entered data to be injected
into the page in an unsafe way, allowing XSS attacks. This code switches
to explicit escaping of strings using escapejs, removing the XSS vector.

r=jgmize