bug 1277577 - Rearrange and document requirements
Add a comment summarizing how each requirement is used in Kuma. For
requirements that are unused, or have a more commonly used variant, add
a note to remove it. Move requirements of requirements to
constraints.txt. Add a README.rst for future devs.

The versions have not been updated. That will be done slowly, in
separate PRs.