bug 1218563: Disable pipeline in Vagrant dev
Previously, the django-pipeline settings PIPELINE_ENABLED and
PIPELINE_COLLECTOR_ENABLED were unset, and were set to default values at
runtime (Off for DEBUG=True, on for DEBUG=False in production).

The Docker configuration sets them explictly, either from environment
variables or the documented defaults.

The Vagrant dev environment starts with DEBUG=False, and then resets it
to DEBUG=True in settings/local, but not until the pipeline is enabled
for production.

This code updates the Vagrant settings that get set to production
defaults, and adds a note that local.py has a short life remaining.