Fix Bug 1149205: Break long titles intelligently.
Previously we've used the CSS property hyphens to break up long titles on short screens but it's not cross browser and it breaks stuff that doesn't need to be broken, obscuring meaning.

This small script targets only long article titles and adds a <wbr> tag to give the browser more accurate direction for where to break titles.

Testing pages:
https://developer-local.allizom.org/en-US/docs/Web/API/MediaStreamAudioDestinationNode/stream

https://developer-local.allizom.org/en-US/Add-ons/Firefox_for_Android/API/window.NativeWindow/window.NativeWindow.contextmenus/window.NativeWindow.contextmenus.remove

https://developer-local.allizom.org/en-US/docs/Web/CSS/:-moz-system-metric%28scrollbar-thumb-proportional%29

(make sure code escaping is working properly)
https://developer-local.allizom.org/en-US/docs/Web/HTML/Element/blockquote