bug 1314290: Fix ALLOWED_HOSTS for api in Docker
Django 1.8.16 checks ALLOWED_HOSTS even when DEBUG=True. This prevented
the api container from accepting connections from the kumascript
container, even though it allowed direct access with http://localhost:8001.

Added 'api' and 'web' as expected hostnames, and the other local names
you get with the default ALLOWED_HOSTS.  Django strips the port number
from the request host, so 'localhost:8000' is always checked as
'localhost', and does not need to appear in the settings.