bug 930102: Add "Hidden When Reading" style
Add "Hidden When Reading" option to style dropdown in editor. In the
menu, it has a similar appearance to when hidden content is displayed in
the editor (that is, with a dotted outline around it). The small black
area at the top-left is where the "hidden" tab is in the edit box;
shobson says this is not something we can easily fix and it isn't a
significant issue anyway, so it should be fine as it is.