Fix Bug 1329743: Compat bugs.
Can't fix the Safari problem. https://bugs.webkit.org/show_bug.cgi?id=166870

The IE problem is with an indicator that was a bit of a hack to begin with - we're styling based on inline styling. IE is changing the inline style from a hex to an rgb colour. So now we have a selector for the rgb value as well.

Text styling difference has to do with differing support for text-overflow: ellipsis. I don't think we want ellipsis here, since it could hide important information, so I removed it.

IE also has no default saying for abbr, so added that.