bug 1271509: Fix past revision clearing doc.html
When a new revision is created, even w/ Revision.objects.get_or_create,
the Revision.save() method is called. This surpised me, I thought it was
skipped, which is only true for bulk creation.

If is_approved=True (the default), the save() method makes the revision
the current revision, which sets Document.html to the revision.content
(default empty string).

I've fixed the storage code to save the initial revision with
is_approved=False, so that we don't accidentally clear doc.html when
adding a historical revision