bug 1271509: Handle doc.current_rev is not latest
At the end of document scraping, chech that the document has a
current_revision. If it doesn't, scrape more revisions and/or more
history, until we find a current revision or run out of revisions.

Page moves mark the page move in the most recent revision, but leave the
n-1 revision as the current_revision.  The default for ./scrape_document
and others is to just scrape one document, leaving these stuck with no
content.