bug 1269104: Optimize DB calls on rev history page
Minimize database queries for the document, revisions, and creators,
to only request data used in template rendering, and to reduce the total
number of DB queries from 15 to 10. This gives a 2.5x speed-up in the
local environment, and gets us a step closer to allowing scrapers.

New template context also adds "Previous" link for first translation
revision, when the English base revision is known, for a consistent UI.