Updates for iframe display behaviour (#4379)
* Fixes display problems with iframe and editors

Swap from vertical to horizontal occurs at small-desktop-and-up media query now, in all cases.

To not have to continuously fight with heights, I defined a max-height. The maximum height allows for ~6 lines of examples, which is one more than our "rule" of 5. If the specific example goes beyond this, then the examples area aka left side of the editor, will be scrollable to reveal the additional examples.

There is also a PR on the interactive editor side that I will have to open and merge so all the changes, and behaviours are evident.