All servers get blue favicon
Previously we had different coloured favicons for different servers so
we could make changes on dev and staging fearlessly. Since the move to
AWS the red favicon has been displaying on all servers. This is bad
because it's the colour of "it's safe to recklessly make changes".
Until https://github.com/mozmeao/infra/issues/585 is fixed all servers
should display the "Be careful!" colour.