bug 1399639: Configure python for node-gyp
node-gyp is the compiler for binary node.js packages, needed to compile
the optional New Relic native-metrics package. node-gyp requires Python,
and looks for it at /usr/bin/python by default. This location doesn't
exists, because Ubuntu is transitioning to Python 3 and wants users to
be explicit about Python 2 versus Python 3. This PR sets the Python path
to the installed Python 2.7 binary.