bug 1431205: Expect less search results in dev
The sample database has about 1000 docs, and all the 'CSS' results are
returned by the default search filter. Production has more than 50K
docs, and some docs with 'CSS' are not in the default filter. Use the
search engine's doc count to try to detect the sample DB and change test
behaviour.