bug 1402950: XFAIL dashboard page2 test in Firefox
The second page of revisions is loaded with AJAX. About 50% of the time,
this confuses Firefox 57 and Selenium 3.8.1. A timeout error occurs when
looking for the page to switch to 2, but the screenshot shows it has. On
the second retry, Selenium fails to establish a connection with Firefox.
This doesn't happen with Chrome 63.

On the first timeout w/ Firefox, mark the test as XFAIL.