bug 1316610: Switch ALLOW_ROBOTS to lists
Instead of a yes/no setting, use lists of hostnames that allow robots,
and deny robots for other hostnames.

For the main website, robots.txt allows robots, but forbids some paths
with dynamic or user-specific content.

For the untrusted attachments and samples domain, allow robots.
Previously, they were not allowed in AWS, but they seem to be an
important part of the content.