bug 1399639: Update form testing for user profile
In a browser, a selected checkbox submits the value "on", and is omitted
from the submitted form when off. pyquery 1.2.15 changes to match jQuery
and browser behaviour for .val(). Instead, use .checked to determine if
the checkbox is selected. Note: Django accepts falsey values "False",
"0", etc. as well, to signify that a checkbox is unchecked.

Email is not an editable text field on the profile edit page. The user
picks their email or adds a new unverified email on a different page.
The profile edit form ignores the submitted email, so it is removed from
tests.