bug 1401246: Remove django-localdevstorage
This library looked for missing local files on a remote server,
downloading them as needed. This was useful when loading a production
database on a development machine. However, we no longer SERVE_MEDIA in
production, so a local development box can't access the raw URLs. In any
case, the sample database is preferred over a production database.

This removes the library and customized storage backend, basically
reverting commit 5042d297.