bug 1431259: Don't cache the create page redirect
Use a "no-cache" cache control header when redirecting to create a page,
to avoid a loop where you have to wait for the cache to expire to see
the new page. This also updates the shared_cache_control decorator to
take no action if it detects the cache control header is set to
no-cache.