bug 1429933: Use Django's URL locale framework
Base Kuma's LocaleMiddleware on Django's LocaleMiddleware, and update
tools to be closer to Django's default localization framework.

One large changes is that urlpatterns are split into ones with a
language prefix, and ones without. Django 1.8 still requires the
LANGUAGE_URL_IGNORED_PATHS hack, but Django 1.11 can better determine if
a given URL should have a locale prefix.

Django's language codes are lowercase (en-us), while Kuma's language
codes use uppercase for the region (en-US). In order to keep the same
URLs, several changes are required, such as replacing
LocaleRegexURLResolver and i18n_patterns.

Kuma's previous locale tools (reverse, Jinja's wiki_url) defaulted to no
locale prefix. With the new framework, the default is the currently
active locale, and it is retained in request.path_info. These changes
require some adjustments, such as in the DocumentZone middleware.

Some Kuma behaviours are not part of the Django framework. The ?lang
querystring handling is moved to a new LangSelectorMiddleware, and
redirects to non-canonical locales (such as /en/ to /en-US/) would be
404s with Django, but are handled in the LocaleStandardizerMiddleware.
Some other differences, such as removing the Vary: Accept-Language
header, are handled in the overridden LocaleMiddleware class.