bug 1409423: Remove invalid test of banned users
Remove a test that the revision dashboard is still available if you are
logged in as a banned user. This is only possible with the test client.
On the site, banned users are logged out, and not allowed to log in again.
The test is testing a situation that is no longer possible.