bug 1461350: Default to locale=en-US in tests
When using Kuma's reverse, don't pass locale="en-US" (common in tests to
avoid a redirect) or locale=request.LANGUAGE_CODE (used in app code to
avoid a redirect), but instead rely on the default behavior of
including the active language.

Continue passing the locale parameter for non-default locales, or if the
test makes a request that sets the active language (usually by making a
request to a non-default locale).

The active language is reset to en-US at the start of each test with an
always-run fixture set_default_language.