bug 1461350: Improve is_valid_path for 404s
The fallback view mindtouch_to_kuma_redirect matches everything, so
using the resolver to detect invalid paths doesn't work. Refactor so we
can detect if mindtouch_to_kuma_redirect would return a 404, and avoid
the redirect.

This also means that RemoveSlashMiddleware works for non-locale URLs.
Coverage is improved, but (as noted) not all code paths are exercised by
tests, and some appear to be impossible.