bug 1461209: Ignore DisallowedHost exception
The DisallowedHost exception is raised when a client's requested
hostname does not match one of the accepted hostnames in ALLOWED_HOSTS.
This is common in AWS from security scanners searching AWS IP blocks, or
when AWS allocates an IP address or EC2 instance that was owned by a
previous website. There is no security risk or positive action that can
be taken, so it should not result in New Relic alerts.