bug 1401246: Avoid import error from configparser
configparser 3.5.0 installs a backports directory without an __init__.py
file, which is compatible with Python 3 but not Python 2. This results
in a repeated warning (with PYTHONWARNINGS=all):

/usr/local/lib/python2.7/pkgutil.py:183: ImportWarning: Not importing directory '/usr/local/lib/python2.7/site-packages/backports': missing __init__.py

This is tracked upstream at:

https://bitbucket.org/ambv/configparser/issues/20/importwarning-missing-__init__py

Touching the file creates it if it doesn't exists, which avoids the
warning, and should still work when configparser fixes the issue.