bug 1461350: Update URLs for attachment host
Update the URLs for the attachment host to ensure that code samples and
file attachments are processed with locale prefixes.

This adds a test that confirms that a code sample can be viewed with
ENABLED_RESTRICTIONS_BY_HOST. This test fails before this change. Other
tests are split up so they test a single configuration.

The urls cache is set when the RestrictedEndpointsMiddleware overrides
the urlconf. Django's TestCase resets this cache, but pytest doesn't
unless mark.urls is used. mark.urls is set to the default urlconf, the
test sets it to the restricted urlconf, and pytest-django cleans it up.