bug 1401246: Fix test for raw file with 304
When the ConditionalGetMiddleware in 1.9 and 1.10 detects that a file
attachment is unchanged, it returns a 304 Not Modified response that
removes caching headers and the Last-Modified header. The test is
xfailed for these Django versions before the header check.

Django 1.11 retains the headers, so tests will pass like 1.9. The tests
are modified to add the attachment host to ALLOWED_HOSTS, which is
checked in tests starting in 1.10.