bug 1401246: Avoid undefined URLs in settings
The AdminEmailHandler in Django 1.11 includes a copy of the settings in
the email.  With a minimal urlpatterns (like in a test and for the
production attachments host), this results in a runtime NoReverseMatch
exception when building the email.

To fix this issue, the settings LOGIN_URL and LOGIN_REDIRECT_URL change
to the pattern names, which is allowed by Django. As noted in the 1.10
release notes, LOGOUT_URL hasn't been used from before 1.0, and is
removed.