bug 1189160: Clean waffle cache in tests
django-waffle 0.12 and later include a more aggressive cache for flags
and switches. When they are created as database objects in tests, they
can be cached, and not cleared when the test transaction is rolled back.

When possible and straightforward, use waffle.testutils' override_flag
and override_switch, to temporarily set the flag or switch, without a
database write.

When not possible, such as when the flag is set in a setUp method in a
TestCase base class, ensure that the .delete() method is called, so that
the waffle cache will be cleared.