bug 1434690: Redirect favicon.ico to domain PNG
The favicon is set in the HTML header, so that production, staging, and
local development will get different color icons. This adds the same
logic to the redirect for /favicon.ico, which is used when the document
is not HTML, like /media/revision.txt.

/favicon.ico is also used by older browsers like IE 10, which doesn't
support PNG, however we've dropped support for these older browsers.