bug 1468491: Install tools with package.json
In the docker image, use "npm install" to install the versions in
package.json to /tools, and symlink the binaries to /usr/local/bin.
The node packages used in build process are in dependencies, and
installed with --production.

npm 5, bundled with node 8, adds package-lock.json as an improved
npm-shrinkwrap.json.

To make it easier to update npm pacakges, the user is set to 1000 (kuma
in the container), and folders and files are owned by kuma. There is a
new --chown parameter to COPY in Dockerfile, and this is used to avoid
having to chown files after copying.

In the TravisCI builds, install tools using "npm install". Rename
INSTALL_PIPELINE to INSTALL_NPM_TOOLS to reflect new usage.

The tox stylelint target is renamed "assetlint", so that we can move
more style checks to TravisCI.