bug 1468491: Update front end dev docs
Rewrite front-end development sections to add an overview of asset
processing in development and production modes, the limitations of host
processing with gulp, and how to run linters. Mark the gulp installation
as optional.

Add a Makefile command and instructions for regenerating
package-lock.json.

Change section titles to MDN standard of sentence case.