bug 1477531: Avoid domain change in host redirect
If the path contains a full URL, then a naive urljoin will use the
scheme and hostname from the path, allowing open redirects on the legacy
domains. Instead, use urlsplit / urlunsplit to create a URL. This will
not affect "regular" paths", but will modify paths that look like full
URLs.

The issue with urljoin is described in the Python docs:

https://docs.python.org/3/library/urllib.parse.html#urllib.parse.urljoin