bug 1180144: Use exact matches for keywords
For the keyword queries (kumascript_macros, css_classnames,
html_attributes) match on the exact value provided, rather than add in a
prefix match by default.

If a '*' or '?' is included, continue with an exact match without those
characters, and a wildcard match with the additional characters.

A keyword query of just '*' or '?' is ignored.