bug 1464093: Update ElasticSearch docs for 5.6.10
Kuma is using 5.6.10 in development and production. The previous
production version was 2.4, but the docs were not updated.

Add some more details about the ElasticSearch support policy, and which
version Kuma is targeting. The version history was removed from
Wikipedia, and there doesn't seem to be a good release history on the ES
website, so link to the end-of-life support page.