bug 1481078: Fix zone redirects with format chars
Previously, the sub-path was added to the destination URL by appending
it in a callable. If the sub_path included a Python format string, like
/foo/bar/{var}, then django-redirect-urls would run .format on it, and
raise a KeyError for {var}.

Instead of a callable, use a format string like
'/en-US/docs/Learn/{sub_path}'. This way, django-redirect-urls will
format the string with the sub_path captured in the regex.