bug 1380388: Refactor deleted document tests
Convert Django tests in test_views.py to pytest tests in
test_document_views.py. Expand the tests to test the new template
features - full deletion logs and a purge button.

This also move delete_client up one level and renames it to
moderator_client. This isn't strictly needed, but is useful for
follow-on wiki tests.