bug 1380388: Assert perms, deletion log template
The template deletion_log.html contains details of who deleted a
document, and is restricted to moderators who could restore (and
possibly purge) the document in the view. Add a jinja2 'assert' function
that raises an error if rendered without this permission, and then
remove the branches that check the permission.

This required refactoring permissions to create an independent moderator
user and client.