bug 1477016: Update to Django 1.11 i18n logic
Update our forked translation code from Django 1.8 to 1.11 logic. The
main change is that in a few places a language code of None instead uses
settings.LANGUAGE_CODE (en-US), so that URLs will get the default when
no language is active.

Add some inline comments, prefixed with "# Kuma:" making it clearer
where we've added or removed logic from the Django code.

Django 1.10 adds a "prefix_default_language" setting, which allows
omitting the default language in the path, as an aid for converting to
language-prefixed URLs. The forked code assumes it is True (the default
locale must appear in the path), and removes 1.11 logic to flexibly
handle this new setting.

https://docs.djangoproject.com/en/1.11/topics/i18n/translation/#django.conf.urls.i18n.i18n_patterns