bug 1490727: Force English for payments form
Force English for the payments form, when it appears as a banner or as
part of the permanent page. Also force the thank-you text to English.
Allow the FAQ and the support form to be translated.

This allows Pontoon contributors to prepare translations, so that they
will be ready when we ship non-English locales.