bug 957802: Update docs to use docker-compose exec.
docker-compose 1.23.0 changed the way containers are named to add a
random string of hexadecimal digits to each name, so the instructions
in docs/installation.rst and in docs/development.rst that use the
`docker exec -i <container_name>` form no longer work since the
container name is not predictable.

This patch updates those documents to use `docker-compose exec <service>`
instead.

The patch also updates requirements/docs.txt to keep the version of
Jinja2 in sync with what is in requires/default.txt

Test Plan:
- `cd docs; make html` and verify that the generated file looks right.
- install a new copy of Kuma, following these revised instructions