bug 1487403: upgrade cleancss to version 4.x (#5135)
* bug 1487403: upgrade cleancss to version 4.x

This patch updates package.json to upgrade from clean-css 3.x
to clean-css-cli 4.x. It makes other modifications to add and
correctly process the command-line arguments that are required
for the new version of cleancss.

Test plan:
- spot check the files in static/build/styles to make sure they
  have not changed too dramatically. In particular, pay attention
  to relative paths for images and fonts.
- run locally to verify that there is no obvious CSS breakage.

* I found a bug in one of the cleancss -O2 optimizations and
the package is not very actively maintained right now, so
playing it safe even though this makes our CSS files a little
bigger than they were.