Issue 683: Kuma-side changes for simplifying KumaScript
For https://github.com/mdn/sprints/issues/683 I'm trying to simplify
KumaScript. My first target is to remove the layers of HTTP caching
between Kuma and KumaScript so that I can try to remove the memcached
dependency (and save on hosting costs).

This patch is changes on the Kuma side to prepare for that refactoring.
First, the patch includes a couple of new scripts that I used to build
a test corpus of raw and rendered documents so I can be sure that my
KumaScript changes do not actually change how documents are rendered.

More substantially, this patch modifies kuma/wiki/kumascript.py to
change the get() function to actually make a POST request. This means
that the KumaScript no longer needs to support the GET request for
rendering