Bug 1509357: redirect En/ and EN/ to en-US/
Kuma's locale middleware correctly detects the desired language when
non-standard capitalization is used. "En" is detected as "en-US",
"PT" is detected as "pt-PT", and "FR-FR" is detected as "fr". But even
though the language is corrected detected from the path, the
LocaleStandardizerMiddleware does a case-insensitive prefix match
between the path and the detected language and responds with a 404
in this case becasue "En" is not a prefix of "en-US" and because
"fr" is not a prefix of "FR-FR".

This patch modifies LocaleStandardizerMiddleware to do case-insensitive
comparisons so that locales like "En", "EN", "PT" and "FR-FR" that
previously gave unnecessary 404s now redirect to a canonical locale.

Note that the middleware has all along allowed locales like "en-U" to
redirect to "en-US". It probably shouldn't do that, but it does, so this
new addition of case-insensitive matching now means that "en-u" also
redirects to "en-US".

Test Plan:

Verify that the following URLs redirect to valid pages:

- https://developer.mozilla.org/En/docs/Web/CSS
- https://developer.mozilla.org/EN/docs/Web/CSS
- https://developer.mozilla.org/PT/docs/Web/CSS
- https://developer.mozilla.org/fr-FR/docs/Web/CSS