Bug 1509430: rewrite and simplify KumaScript
This is a big patch that effectively rewrites the KumaScript server
to use async/await instead of fibers and Future and hirelings. It
also removes unused features and dependencies and modernizes the code.

It removes the need for a memcached server, and renders documents much
faster than the old version of the server does.

- environment.js, cache.js, config.js, plus the macros directory

- Start using prettier prettier

- Add await to lots of macros

- more tests

- Add a test for the template() function

- Add MacroSyntaxError class

- update pegjs version to get better token location reporting

- better error reporting and handling

- Port over the server.js file

- add server tests

- add a parallel request server test

- containerize ks2

- Kuma-side changes for simplifying KumaScript

For https://github.com/mdn/sprints/issues/683 I'm trying to simplify
KumaScript. My first target is to remove the layers of HTTP caching
between Kuma and KumaScript so that I can try to remove the memcached
dependency (and save on hosting costs).

This patch is changes on the Kuma side to prepare for that refactoring.
First, the patch includes a couple of new scripts that I used to build
a test corpus of raw and rendered documents so I can be sure that my
KumaScript changes do not actually change how documents are rendered.

More substantially, this patch modifies kuma/wiki/kumascript.py to
change the get() function to actually make a POST request. This means
that the KumaScript no longer needs to support the GET request for
rendering

- kumascript environment tweaks and macro debugging, including string.asyncReplace() function

- Change the way the macro execution envionment is bound together; add a test to verify that all macros compile; fix the macros that do not compile.

- fix a bug where the Compat macro was writing into the bdd data

- Fix final bugs to reach rendering parity

- Better error classes + simplify requestHTTPResource

- tweak the cache size