Fix the label of the seoSummary style to be just "Summary" (#5184)
In PR #5154, the name of the seoSummary class was changed
in the CKEditor menu to just "Summary". This patch makes
the same change to the CSS that is used to display the
style when in the editor, so that the label drawn above
the styled content says that instead of "SEO Summary".