Bug 1518844: convert bugzilla links to github links
This is a followup patch for https://github.com/mozilla/kuma/pull/5183
That original PR changed the content bug reporting link in the
navigation header, but missed it in the 404 page and in a few
other places. This patch removes all instances of "form.doc" from
the Kuma source code.