bug 1520574: Reduce file data passed to kumascript
With node 10.14.0, there is an 8K limit to HTTP headers. When file data
is base64 encoded as a header, it can grow beyond 8K for some pages with
10 or more attachments. This results in a 400 error when rendering.

Reduce the attachment metadata to the filename and URL, which are the
two attributes used by Embed_text and EmbedSVG, the only macros that use
this data. Also, drop the "attachments" header, which encodes the same
data but is unused.

This is a first step to removing these macros and no longer passing
attachment data in the default environment.