Bug 1520616: require login for the revision dashboard
We believe that scrapers/crawlers hitting the revision dashboard
(while ignoring our robots.txt file) are causing significant
database load that is reducing performance of the site.

This patch changes the dashboards/revisions endpoint to require
login and to not be cached. The changes to kuma/dashboards/views.py
are trivial; most of the changes are adjustments to tests.