bug 1490727: Move Stripe error handling to view
If we request a customer that doesn't exist, a InvalidRequestError will
be raised. Move the handling of this from the utility function to the
view that uses it, so that the username and email don't have to be
passed to the function just for logging.

This also switches to log.exception, which should print the traceback to
the log, and uses the message / args format supported by Django logging.

The code now handles all Stripe errors the same way, but doesn't have a
generic exception handler. I find that generic handlers hide errors and
slow down debugging.