bug 1490727: Unsubscribe User in pre_delete signal
When deleting a User with a stripe_customer_id, attempt to cancel any
subscriptions first. If this fails (unknown customer, other Stripe API
issue), an exception will raise and the user deletion will stop, so that
an admin can attempt to manually delete the subscription before the User
data is lost.