Bug 1518844, ensure focus is clearly indicated on amount radio buttons (#5197)
Problem is described here(although not very clearly ;p)
https://github.com/mdn/sprints/issues/713

In essence, the problem is that even though the radio buttons do receive focus, and is navigable using the arrow keys, it is not clear when they have focus.

Here is the current behaviour on production. As can be seen, when focus is moved from the email field to the radio buttons, nothing actually indicates that the focus has moved to the `checked` radio button. It feels like focus moves from the email field, to someplace on the form, and then to the custom amount field.

![radio-fields-focus-current](https://user-images.githubusercontent.com/10350960/51323089-dc3a2180-1a6f-11e9-8bb1-7de5a3e942b9.gif)

Adding the `focus-visible` class with the clear focus ring solves the problem

![radio-fields-focus](https://user-images.githubusercontent.com/10350960/51322758-00493300-1a6f-11e9-8020-e0a97026d472.gif)

@davidflanagan r?