Bug 1522718: render stale documents in series, not in parallel
Kuma's render_stale_documents task splits the list of stale
documents into groups of 5, then re-renders each group in
parallel. This typically causes about 12 requests to the KumaScript
server at the same time, causing load spikes that only occur
once an hour when this task runs.

This patch changes the task so that the subtasks for each group
of documents are chained and run in series instead of in parallel.
This should reduce the load on KumaScript and enable us to drop
down to a single KumaScript server.