bug 957802: Change to "triaged" and "bleached"
escattone did a lot of this work six months ago, and posted his findings
to Discourse:

https://discourse.mozilla.org/t/automating-page-changes-with-the-put-api/29842/8

This switches "Cleaned content" to "Bleached content", and "Re-edit
content" to "Triaged content", to be consistant with that naming scheme.

I'm adding one more content section, and will update the diagrams in
that commit.