bug 1380419: Document static assets
This is an attempt to document how assets are processed in development
and production. My goal was to understand the system enough that we
can safely remove parts that are no longer needed.

I thought assets were complex before, but they are even more complex
than I thought. This is reflected in a long document that jumps back
and forth. Still, I think some changes, like removing django-pipeline
and avoiding collecting assets not served to users, can simplify the
system and increase performance.