bug 1522718: Remove cache lock code
Simplify batch processing by removing the cache lock code. Code coverage
also increases since there are no tests for this code.

The cache lock code was added when rendering stale documents was
expensive, and was being groomed to be the mechanism for mass document
regeneration. With the changes of the past 5 years, this is no longer
the case, and the risk of parallel re-render threads is less.

The cache was also used for feed processing, which could also be
expensive. We currently only display the Hacks Blog feed, and do not
intend to republish more feeds.