bug 1490356: Expand sourcecompletion log message
The "9" in this log line was unclear:

DEBUG: 9:36/55 Source document_current:/pt-BR/docs/Aprender complete, freshness=Fresh, with 0 dependent sources.

This makes it clearer that we're in the 9th round of gathering sources,
on source 36 of 55 for the round:

DEBUG: Round 9, Source 36 of 55: document_current:/pt-BR/docs/Aprender complete, freshness=Fresh, with 0 dependent sources.